♻️Refactor: chip 모집/공개여부에 따른 label 컴포넌트 위치 변경
